backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pydantic[email]==2.5.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
fastapi-users[sqlmodel]==12.1.3
fastapi-users-db-sqlmodel==0.1.0
PyJWT==2.8.0
bcrypt==4.1.2 